+++ cmd/swarmctl/job/cmd.go
@@ -13,7 +13,7 @@ var (
 func init() {
 	Cmd.AddCommand(
 		inspectCmd,
-		lsCmd,
+		listCmd,
 		createCmd,
 		updateCmd,
 		removeCmd,
