+++ cmd/swarmctl/task/cmd.go
@@ -12,7 +12,7 @@ var (
 
 func init() {
 	Cmd.AddCommand(
-		lsCmd,
+		listCmd,
 		inspectCmd,
 		removeCmd,
 	)
